# microservice/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  FeedItem:
    properties:
      action:
        type: string
      create_time:
        type: string
      id:
        type: integer
      show_divider:
        description: 分割线
        type: boolean
      source:
        $ref: '#/definitions/Source'
      user:
        $ref: '#/definitions/User'
    type: object
  FeedListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/FeedItem'
        type: array
    type: object
...
  Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  Source:
    properties:
      id:
        type: integer
      name:
        type: string
      type_name:
        type: string
    type: object
...
  User:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  chat.Message:
    properties:
      content:
        type: string
      sender:
        type: integer
      time:
        type: string
      type_name:
        type: string
    type: object
  chat.UserStatus:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  comment.Comment:
    properties:
      be_replied_content:
        type: string
      be_replied_user_id:
        type: integer
      content:
        type: string
      create_time:
        type: string
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      father_id:
        type: integer
      id:
        type: integer
      img_url:
        type: string
      is_liked:
        type: boolean
      like_num:
        type: integer
      type_name:
        description: first-level -> 一级评论; second-level -> 其它级
        type: string
    type: object
...
  post.Comment:
    properties:
      be_replied_content:
        type: string
      be_replied_id:
        type: integer
      be_replied_user_id:
        type: integer
      be_replied_user_name:
        type: string
      content:
        type: string
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      id:
        type: integer
      is_liked:
        type: boolean
      like_num:
        type: integer
      time:
        type: string
    type: object
...
  post.GetPostResponse:
    properties:
      category:
        type: string
      collection_num:
        type: integer
      comment_num:
        type: integer
      compiled_content:
        type: string
      content:
        type: string
      content_type:
        description: md or rtf
        type: string
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      id:
        type: integer
      is_collection:
        type: boolean
      is_liked:
        type: boolean
      like_num:
        type: integer
      sub_posts:
        items:
          $ref: '#/definitions/post.SubPost'
        type: array
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      time:
        type: string
      title:
        type: string
    type: object
...
  post.PostPartInfo:
    properties:
      category:
        type: string
      collection_num:
        type: integer
      comment_num:
        type: integer
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      id:
        type: integer
      like_num:
        type: integer
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      time:
        type: string
      title:
        type: string
    type: object
  post.PostPartInfoResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/post.PostPartInfo'
        type: array
    type: object
...
  post.SubPost:
    properties:
      comment_num:
        type: integer
      comments:
        items:
          $ref: '#/definitions/post.Comment'
        type: array
      content:
        type: string
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      id:
        type: integer
      img_url:
        type: string
      is_liked:
        type: boolean
      like_num:
        type: integer
      time:
        type: string
    type: object
...
host: forum.muxixyz.com
info:
  contact: {}
  description: The gateway of forum
  title: github.com/Muxi-X/forum-be/microservice/gateway
  version: "1.0"
paths:
...
  /chat/history/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chat.Message'
            type: array
      summary: 获取该用户的聊天记录
      tags:
      - chat
  /chat/userList:
    get:
      description: 获取该用户的聊天列表，包括每个聊天对象的名字和图片
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户列表
          schema:
            items:
              $ref: '#/definitions/chat.UserStatus'
            type: array
      summary: 获取该用户的聊天列表
      tags:
      - chat
...
  /collection/list/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostPartInfoResponse'
      summary: list收藏 api
      tags:
      - collection
...
  /comment/{comment_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: comment_id
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 删除评论 api
      tags:
      - comment
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: comment_id
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
      summary: 获取评论 api
      tags:
      - comment
  /feed/list/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/FeedListResponse'
      summary: list 用户的动态 api
      tags:
      - feed
...
  /post/{post_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: post_id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 删除帖子 api
      tags:
      - post
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: post_id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.GetPostResponse'
      summary: 获取帖子 api
      tags:
      - post
...
swagger: "2.0"
...